Enable C++11 for armcc
ested on our release repositores, and after a few changes in a
couple of repos, everything seems to compile fine.
NOTE: it looks like the the `--cpp11` argument was introduced first
in armcc 5.05, so I don't expect this to work with armcc 5.03 and
lower. Because of this, I wouldn't merge this patch yet, until we
understand all the implications of this change.